+++ ProjectRPG/GameServer/Game/Room/GameRoom.cs
@@ -251,5 +251,13 @@ namespace GameServer.Game
 
             return zones.ToList();
         }
+
+        private void BroadcastAll(IMessage packet)
+        {
+            foreach (var player in _players.Values)
+            {
+                player.Session.Send(packet);
+            }
+        }
     }
 }
